layer2: Added manual mlPeers configuration
This PR adds a flag to manually configure mlPeers to be able to run
metallb-speaker out-of-cluster.

The rational behind this. Imagine a black box (or rack) with two
switches (cumulus) and a couple of kubernetes nodes. The switches are
not part of the kubernetes cluster. Each node connects to both switches
and bgp announce a node IP, as well as service, LB and pods ips to the
switches. This setup tolarates losing one switch without affecting any
functionality.

The switches have (in some setups) a l2-only connection to a customer
network each, with no way to bgp peer to the customer network. We
now want to provide services to the customer in an HA way, this is where
metallb-speaker in l2 mode comes into play. Until now, we used
metallb-speaker 0.6.2, with kubernetes leader election, so that always
one switch announces the LB addresses. This worked well until kubernetes
1.18, which breaks client-go in the old version of metallb 0.6.2.

Long story short. By adding ml-peers to manually provide static ip
addresses for MemberList (rather than using PodIPs), it is possible
again to run metallb-speaker standalone on the switches again. Even
better, because how the announcement works, load might be splitted
across both switches (what was not possible with bare leader election).